Refactor flag split helpers
