+++ packages/cli/src/util/flags/can-prompt.ts
@@ -1,0 +1,5 @@
+import type Client from '../client';
+
+export function canPrompt(client: Client): boolean {
+  return Boolean(client.stdin.isTTY) && !client.nonInteractive;
+}
