+++ packages/cli/test/unit/commands/flags/split.test.ts
@@ -8,89 +8,7 @@ import { defaultProject, useProject } from '../../../mocks/project';
 import { useTeams } from '../../../mocks/team';
 import { useUser } from '../../../mocks/user';
 import type { Flag } from '../../../../src/util/flags/types';
-
-function createTestFlags(): Flag[] {
-  return [
-    {
-      id: 'flag_bool123',
-      slug: 'my-feature',
-      description: 'My awesome feature flag',
-      kind: 'boolean',
-      state: 'active',
-      variants: [
-        { id: 'off', value: false, label: 'Off' },
-        { id: 'on', value: true, label: 'On' },
-      ],
-      environments: {
-        production: {
-          active: false,
-          fallthrough: { type: 'variant', variantId: 'off' },
-          pausedOutcome: { type: 'variant', variantId: 'off' },
-          rules: [],
-        },
-        preview: {
-          active: true,
-          fallthrough: { type: 'variant', variantId: 'off' },
-          pausedOutcome: { type: 'variant', variantId: 'off' },
-          rules: [],
-        },
-        development: {
-          active: true,
-          fallthrough: { type: 'variant', variantId: 'on' },
-          pausedOutcome: { type: 'variant', variantId: 'off' },
-          rules: [],
-        },
-      },
-      createdAt: Date.now() - 86400000,
-      updatedAt: Date.now() - 3600000,
-      createdBy: 'user_123',
-      projectId: 'vercel-flags-test',
-      ownerId: 'team_dummy',
-      revision: 1,
-      seed: 12345,
-      typeName: 'flag',
-    },
-    {
-      id: 'flag_string456',
-      slug: 'welcome-message',
-      description: 'A string feature flag',
-      kind: 'string',
-      state: 'active',
-      variants: [
-        { id: 'control', value: 'control', label: 'Control' },
-        { id: 'treatment', value: 'treatment', label: 'Treatment' },
-      ],
-      environments: {
-        production: {
-          active: true,
-          fallthrough: { type: 'variant', variantId: 'control' },
-          pausedOutcome: { type: 'variant', variantId: 'control' },
-          rules: [],
-        },
-        preview: {
-          active: true,
-          fallthrough: { type: 'variant', variantId: 'control' },
-          pausedOutcome: { type: 'variant', variantId: 'control' },
-          rules: [],
-        },
-        development: {
-          active: true,
-          fallthrough: { type: 'variant', variantId: 'control' },
-          pausedOutcome: { type: 'variant', variantId: 'control' },
-          rules: [],
-        },
-      },
-      createdAt: Date.now() - 172800000,
-      updatedAt: Date.now() - 7200000,
-      createdBy: 'user_123',
-      projectId: 'vercel-flags-test',
-      ownerId: 'team_dummy',
-      revision: 2,
-      seed: 67890,
-      typeName: 'flag',
-    },
-  ];
-}
+import { createTestFlags } from './fixtures';
 
 describe('flags split', () => {
   const selectMock = vi.fn();
